% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.product_names.R
\name{extract.product_names}
\alias{extract.product_names}
\title{pulls product names from GsmArena}
\usage{
extract.product_names(filename, sheetname, brandname, brand_url, brand_pages)
}
\arguments{
\item{filename}{name of the excel file.}

\item{sheetname}{name of the sheet of excel file.}

\item{brandname}{name of the brand.}

\item{brand_url}{url of the products second page}

\item{brand_pages}{total number of pages.}
}
\value{
Dataframe
}
\description{
pull product names of a brand from GSMARENA.
}
\examples{
extract.product_names("myfile","sheet1","Apple","https://www.gsmarena.com/apple-phones-f-48-0-p2.php",2)

}
