% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.reviews.R
\name{extract.reviews}
\alias{extract.reviews}
\title{pulls product reviews from GsmArena}
\usage{
extract.reviews(url, n, filename, sheetname)
}
\arguments{
\item{url}{url of the products second page}

\item{n}{total number of review pages.}

\item{filename}{name of the excel file.}

\item{sheetname}{name of the sheet of excel file.}
}
\value{
Dataframe
}
\description{
pull reviews/opinions from GSMARENA and store in excel sheet.
}
\examples{
extract.reviews("https://www.gsmarena.com/vivo_v5-reviews-8430p2.php",3,"reviews","vivo")

}
